backend/requirements.txt
fastapi
uvicorn[standard]
azure-storage-blob
azure-search-documents
openai
python-dotenv
python-multipart
pymupdf
python-pptx
python-docx
openpyxl
azure-functions
